drivers/led/mbi: make shift_amount and instr arguments uint8_t

+++ drivers/led/mbi.c
@@ -141,7 +141,7 @@ __attribute__((weak)) void mbi_init_config(void);
 __attribute__((weak)) void mbi_init_timers(void);
 
 /* Send 'instr' number of DCLK pulses while LE is asserted high. */
-void mbi_send_instruction(int instr) {
+void mbi_send_instruction(uint8_t instr) {
     writePinLow(MBI_LE_PIN);
     mbi_io_wait;
     writePinHigh(MBI_LE_PIN);
@@ -155,7 +155,7 @@ void mbi_send_instruction(int instr) {
 }
 
 /* Transmit `shift_amount` bits of `data` to shift-register. */
-void mbi_shift_data(uint16_t data, int shift_amount) {
+void mbi_shift_data(uint16_t data, uint8_t shift_amount) {
     while (shift_amount-- > 0) {
         mbi_io_wait;
         writePinLow(MBI_DCLK_PIN);
@@ -171,7 +171,7 @@ void mbi_shift_data(uint16_t data, int shift_amount) {
  *
  * Note: Assumes `instr` is less than `shift_amount`.
  */
-void mbi_shift_data_instr(uint16_t data, int shift_amount, int instr) {
+void mbi_shift_data_instr(uint16_t data, uint8_t shift_amount, uint8_t instr) {
     if (instr < shift_amount) {
         writePinLow(MBI_LE_PIN);
         mbi_shift_data(data, shift_amount - instr);
